feat(core/rust): move nrf and smp out of trezor_lib
* moves irq to sys::irq
* moves from_c_str and from_c_array to rtl::util
* moves smp and trezorhal::nrf to io
