+++ core/embed/io/src/nrf.rs
