+++ core/embed/io/src/smp/api.rs
@@ -1,14 +1,16 @@
-use super::{echo, image_info, process_rx_byte, reset, upload};
-use crate::util::from_c_array;
+use rtl::error::unwrap;
+use rtl::util::from_c_array;
 
-#[no_mangle]
+use super::{echo, image_info, process_rx_byte, reset, upload};
+
+#[unsafe(no_mangle)]
 extern "C" fn smp_echo(text: *const cty::c_char, text_len: u8) -> bool {
     let text = unwrap!(unsafe { from_c_array(text, text_len as usize) });
 
     echo::send(text)
 }
 
-#[no_mangle]
+#[unsafe(no_mangle)]
 extern "C" fn smp_reset() {
     reset::send();
 }
@@ -32,7 +34,7 @@ pub struct NrfAppVersion {
 ///
 /// # Returns
 /// `true` if version was successfully retrieved and parsed, `false` otherwise.
-#[no_mangle]
+#[unsafe(no_mangle)]
 extern "C" fn smp_image_version_get(out: *mut NrfAppVersion) -> bool {
     if out.is_null() {
         return false;
@@ -51,7 +53,7 @@ extern "C" fn smp_image_version_get(out: *mut NrfAppVersion) -> bool {
     }
 }
 
-#[no_mangle]
+#[unsafe(no_mangle)]
 extern "C" fn smp_upload_app_image(
     data: *const cty::uint8_t,
     len: cty::size_t,
@@ -64,7 +66,7 @@ extern "C" fn smp_upload_app_image(
     upload::upload_image(data_slice, hash_slice)
 }
 
-#[no_mangle]
+#[unsafe(no_mangle)]
 extern "C" fn smp_process_rx_byte(byte: u8) {
     process_rx_byte(byte)
 }
