+++ core/embed/io/src/smp/base64.rs
