+++ core/embed/io/src/smp/crc16.rs
