+++ core/embed/projects/bootloader/Cargo.toml
@@ -79,7 +79,7 @@ hash_processor = ["sec/hash_processor"]
 iwdg = ["sec/iwdg"]
 lockable_bootloader = ["io/lockable_bootloader"]
 mcu_attestation = ["sec/mcu_attestation"]
-nrf = ["io/nrf", "trezor_lib/nrf"]
+nrf = ["io/nrf"]
 nrf_auth = ["io/nrf_auth"]
 power_manager = ["io/power_manager", "trezor_lib/power_manager", "trezor_lib/pmic"]
 pvd = ["sys/pvd"]
