+++ core/embed/projects/firmware/Cargo.toml
@@ -119,7 +119,7 @@ hash_processor = ["sec/hash_processor"]
 iwdg = ["sec/iwdg"]
 lockable_bootloader = ["io/lockable_bootloader"]
 mcu_attestation = ["sec/mcu_attestation", "upymod/mcu_attestation"]
-nrf = ["io/nrf", "trezor_lib/nrf"]
+nrf = ["io/nrf"]
 nrf_auth = ["io/nrf_auth"]
 optiga = ["sec/optiga", "upymod/optiga", "trezor_lib/optiga"]
 power_manager = [
