+++ core/embed/projects/kernel/Cargo.toml
@@ -81,7 +81,7 @@ secmon_layout = ["models/secmon_layout"]
 secret = ["sec/secret"]
 secure_aes = ["sec/secure_aes"]
 secure_mode = ["io/secure_mode"]
-smp = ["io/smp", "trezor_lib/smp", "trezor_lib/nrf"]
+smp = ["io/smp"]
 suspend = ["io/suspend"]
 tamper = ["sec/tamper"]
 telemetry = ["sec/telemetry"]
