+++ core/embed/projects/prodtest/Cargo.toml
@@ -79,7 +79,7 @@ iwdg = ["sec/iwdg"]
 lockable_bootloader = ["io/lockable_bootloader"]
 mcu_attestation = ["sec/mcu_attestation"]
 nfc = ["io/nfc"]
-nrf = ["io/nrf", "trezor_lib/nrf"]
+nrf = ["io/nrf"]
 nrf_auth = ["io/nrf_auth"]
 optiga = [
     "sec/optiga",
@@ -100,7 +100,7 @@ sdram = ["sys/sdram"]
 secmon_header = []
 secret = ["sec/secret"]
 secure_aes = ["sec/secure_aes"]
-smp = ["io/smp", "trezor_lib/smp"]
+smp = ["io/smp"]
 suspend = ["io/suspend"]
 tamper = ["sec/tamper"]
 telemetry = ["sec/telemetry", "trezor_lib/telemetry"]
