+++ core/embed/rtl/src/util.rs
@@ -85,6 +85,50 @@ impl From<&str> for FatPtr<cty::c_char> {
     }
 }
 
+/// Constructs a string from a C string.
+///
+/// # Safety
+///
+/// The caller is responsible that the pointer is valid, which means that:
+/// (a) it points to a memory containing a valid C string (zero-terminated
+/// sequence of characters), and
+/// (b) that the pointer has appropriate lifetime.
+pub unsafe fn from_c_str<'a>(c_str: *const cty::c_char) -> Option<&'a str> {
+    if c_str.is_null() {
+        return None;
+    }
+    unsafe {
+        let bytes = core::ffi::CStr::from_ptr(c_str as _).to_bytes();
+        if bytes.is_ascii() {
+            Some(core::str::from_utf8_unchecked(bytes))
+        } else {
+            None
+        }
+    }
+}
+
+/// Construct str from a C array.
+///
+/// # Safety
+///
+/// The caller is responsible that the pointer is valid, which means that:
+/// (a) it points to a memory containing array of characters, with length `len`,
+/// and
+/// (b) that the pointer has appropriate lifetime.
+pub unsafe fn from_c_array<'a>(c_str: *const cty::c_char, len: usize) -> Option<&'a str> {
+    if c_str.is_null() {
+        return None;
+    }
+    unsafe {
+        let slice = core::slice::from_raw_parts(c_str as *const u8, len);
+        if slice.is_ascii() {
+            Some(core::str::from_utf8_unchecked(slice))
+        } else {
+            None
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
