+++ core/embed/sys/src/lib.rs
@@ -2,6 +2,7 @@
 
 mod ffi;
 
+pub mod irq;
 #[cfg(feature = "dbg_console")]
 pub mod syslog;
